refactor(webapp): rename RRSet to RecordList and restructure
