+++ webapp/src/components/Field/Record/A.vue
@@ -1,6 +1,6 @@
 <script>
 import { ipAddress } from 'vuelidate/lib/validators';
-import Record from './Record.vue';
+import Record from '../Record.vue';
 
 export default {
   name: 'RecordA',
